Added more ScriptableObjects
- Created ScriptableObjects for player sprites
- Created ScriptableObjects for projectile data and an abstract class for projectile behaviour
- All this should, in theory, allow for more modular code and make it easier to design new projectiles
- Enemy tank turret now rotates during aiming, but the movement is still kind of janky

+++ TYVM Game/Assets/Scripts/Player/Health.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Health : MonoBehaviour {
 
     [SerializeField]
-    private int health = 3;
+    private float health = 3f;
     private GameLogic gameLogic;
 
     private void Awake() {
@@ -16,7 +16,7 @@ public class Health : MonoBehaviour {
 
     }
 
-    public void TakeDamage(int damage) {
+    public void TakeDamage(float damage) {
         health -= damage;
         if (health <= 0) {
             DestroySelf();
